# src/main/java/org/cdb/filesystem/dto/file/ApiError.java
package org.cdb.filesystem.dto.file;

import lombok.AllArgsConstructor;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
public class ApiError
{
    private Integer code;
    private String message;
}