# src/main/java/org/cdb/filesystem/dto/file/ApiFileAddRequest.java
package org.cdb.filesystem.dto.file;

import jakarta.persistence.Lob;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class ApiFileAddRequest extends ApiFileCommon
{
    @Lob
    private String data;
}